Task 9.2 tidied up
