--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/ConfirmationService.java
@@ -1,6 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface ConfirmationService {
-
-    void confirm(User user);
-}
