--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/OrderRepository.java
@@ -1,6 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface OrderRepository {
-
-    void save(OrderRequest order);
-}
