--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/OrderService.java
@@ -1,6 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface OrderService {
-
-    boolean order(final User buyer, final User seller, final Product product);
-}
