--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/PaymentService.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface PaymentService {
-    void pay(OrderRequest order);
-}
