+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/oderprocessing/OrderValidator.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.eshopping.oderprocessing;
 
 public class OrderValidator {
