+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/order/OrderRequestRetriever.java
@@ -1,4 +1,7 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.eshopping.order;
+
+import com.kodilla.good.patterns.challenges.eshopping.product.Product;
+import com.kodilla.good.patterns.challenges.eshopping.user.User;
 
 public class OrderRequestRetriever {
