+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/product/Product.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.eshopping.product;
 
 public class Product {
