+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/services/ConfirmationService.java
@@ -1,0 +1,8 @@
+package com.kodilla.good.patterns.challenges.eshopping.services;
+
+import com.kodilla.good.patterns.challenges.eshopping.user.User;
+
+public interface ConfirmationService {
+
+    void confirm(User user);
+}
