+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/services/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.kodilla.good.patterns.challenges.eshopping.services;
+
+import com.kodilla.good.patterns.challenges.eshopping.order.OrderRequest;
+
+public interface OrderRepository {
+
+    void save(OrderRequest order);
+}
