+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/eshopping/services/PaymentService.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.challenges.eshopping.services;
+
+import com.kodilla.good.patterns.challenges.eshopping.order.OrderRequest;
+
+public interface PaymentService {
+    void pay(OrderRequest order);
+}
